+++ CHANGELOG.md
@@ -17,6 +17,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Theme**) Adjust "primary color" of dynamic color theme on manga pages
 - (**Theme**) Increase max length of custom theme names to 32 (previously 16)
 
+### Fixed
+- (**Library**) Fix total library size chip color in light mode 
+
 ## [20251230.01] (r2937) - 2025-12-30
 
 ### Added